+++ .github/CHANGELOG.md
@@ -7,3 +7,12 @@
 * Link added in side nav to the Impact Calculator
 * Integration of Google Analytics
 * Edits to `angular.json` and `tsconfig` to suppress warnings (`allowedCommonJsDependencies`)
+
+## v0.719.2020
+
+* Added SVG kudos to `explore` component
+* Blocknative
+    * Integrated onclick into notify popups, linking to Etherscan for the tx
+    * Allowed `connectWallet` to retain state in localstorage and when change between components
+    * Allowed users to `changeWallet`, clearing out localstorage
+    * Added Authereum and Dapper as options